autopilot-speed/requirements.txt
fastapi
uvicorn[standard]
pydantic
numpy
scipy
python-can
websockets
paho-mqtt
